move parallel grid to source_base
