+++ source/source_estate/module_charge/charge_extra.h
@@ -5,7 +5,7 @@
 #include "source_cell/unitcell.h"
 #include "source_pw/module_pwdft/structure_factor.h"
 #ifdef __MPI
-#include "source_pw/module_pwdft/parallel_grid.h"
+#include "source_base/parallel_grid.h"
 #endif
 
 /**
