+++ source/source_estate/module_charge/charge_init.cpp
@@ -12,7 +12,7 @@
 #include "source_base/timer.h"
 #include "source_base/tool_threading.h"
 #include "source_cell/magnetism.h"
-#include "source_pw/module_pwdft/parallel_grid.h"
+#include "source_base/parallel_grid.h"
 #include "source_io/module_output/cube_io.h"
 #include "source_io/module_chgpot/rhog_io.h"
 #include "source_io/module_wf/read_wf2rho_pw.h"
