+++ source/source_estate/module_charge/symmetry_rho.h
@@ -3,7 +3,7 @@
 #include "source_basis/module_pw/pw_basis.h"
 #include "source_cell/module_symmetry/symmetry.h"
 #include "source_estate/module_charge/charge.h"
-#include "source_pw/module_pwdft/parallel_grid.h"
+#include "source_base/parallel_grid.h"
 
 class Symmetry_rho
 {
