+++ source/source_hamilt/module_surchem/test/CMakeLists.txt
@@ -16,20 +16,20 @@ AddTest(
   TARGET MODULE_HAMILT_surchem_cal_pseudo  
   LIBS parameter planewave device base psi
   SOURCES cal_pseudo_test.cpp  ../cal_pseudo.cpp ../surchem.cpp ../cal_epsilon.cpp ../../../source_pw/module_pwdft/structure_factor.cpp
-  ../../../source_pw/module_pwdft/parallel_grid.cpp 
+  ../../../source_base/parallel_grid.cpp 
 
 )
 
 AddTest(
   TARGET MODULE_HAMILT_surchem_cal_totn
   LIBS parameter planewave device base
-  SOURCES cal_totn_test.cpp  ../cal_totn.cpp ../surchem.cpp ../../../source_pw/module_pwdft/parallel_grid.cpp 
+  SOURCES cal_totn_test.cpp  ../cal_totn.cpp ../surchem.cpp ../../../source_base/parallel_grid.cpp 
 )
 
 AddTest(
   TARGET MODULE_HAMILT_surchem_cal_vcav
   LIBS parameter planewave device base container
-  SOURCES cal_vcav_test.cpp  ../cal_vcav.cpp ../surchem.cpp ../../../source_pw/module_pwdft/parallel_grid.cpp 
+  SOURCES cal_vcav_test.cpp  ../cal_vcav.cpp ../surchem.cpp ../../../source_base/parallel_grid.cpp 
   ../../module_xc/xc_grad.cpp ../../module_xc/xc_functional.cpp
   ../../module_xc/xc_lda_wrap.cpp ../../module_xc/xc_gga_wrap.cpp
   ../../module_xc/libxc_setup.cpp
@@ -44,7 +44,7 @@ AddTest(
 AddTest(
   TARGET MODULE_HAMILT_surchem_cal_vel
   LIBS parameter planewave device base container
-  SOURCES cal_vel_test.cpp  ../cal_vel.cpp ../surchem.cpp ../cal_epsilon.cpp ../minimize_cg.cpp ../../../source_pw/module_pwdft/parallel_grid.cpp 
+  SOURCES cal_vel_test.cpp  ../cal_vel.cpp ../surchem.cpp ../cal_epsilon.cpp ../minimize_cg.cpp ../../../source_base/parallel_grid.cpp 
   ../../module_xc/xc_grad.cpp ../../module_xc/xc_functional.cpp
   ../../module_xc/xc_lda_wrap.cpp ../../module_xc/xc_gga_wrap.cpp
   ../../module_xc/libxc_setup.cpp
