+++ source/source_hamilt/module_surchem/test/setcell.h
@@ -7,7 +7,7 @@
 #include "source_cell/module_neighbor/sltk_atom_arrange.h"
 #include "source_cell/module_neighbor/sltk_grid_driver.h"
 #include "source_cell/unitcell.h"
-#include "source_pw/module_pwdft/parallel_grid.h"
+#include "source_base/parallel_grid.h"
 #include "source_pw/module_pwdft/structure_factor.h"
 
 namespace GlobalC
