+++ source/source_io/module_chgpot/get_pchg_lcao.h
@@ -3,7 +3,7 @@
 
 #include "source_cell/klist.h"
 #include "source_estate/module_dm/density_matrix.h"
-#include "source_pw/module_pwdft/parallel_grid.h"
+#include "source_base/parallel_grid.h"
 #include "source_basis/module_pw/pw_basis.h"
 #include "source_psi/psi.h"
