+++ source/source_io/module_output/write_cube.cpp
@@ -1,6 +1,6 @@
 #include "source_base/element_name.h"
 #include "source_base/parallel_comm.h"
-#include "source_pw/module_pwdft/parallel_grid.h"
+#include "source_base/parallel_grid.h"
 #include "source_io/module_output/cube_io.h"
 
 #include <vector>
