+++ source/source_io/test/read_wf2rho_pw_test.cpp
@@ -9,7 +9,7 @@
 #include "source_estate/module_charge/charge.h"
 #include "source_estate/module_charge/symmetry_rho.h"
 #include "source_hamilt/module_xc/xc_functional.h"
-#include "source_pw/module_pwdft/parallel_grid.h"
+#include "source_base/parallel_grid.h"
 #include "source_io/module_wf/read_wf2rho_pw.h"
 #include "source_io/module_wf/write_wfc_pw.h"
 #include "source_base/module_out/filename.h" // mohan add 2025-05-17
