+++ source/source_pw/module_pwdft/CMakeLists.txt
@@ -42,7 +42,6 @@ list(APPEND objects
     structure_factor.cpp
     structure_factor_k.cpp
     soc.cpp
-    parallel_grid.cpp
     elecond.cpp
     fs_nonlocal_tools.cpp
     fs_kin_tools.cpp
