+++ source/source_pw/module_pwdft/structure_factor.h
@@ -4,7 +4,7 @@
 #include "source_base/complexmatrix.h"
 #include "source_basis/module_pw/pw_basis_k.h"
 #include "source_cell/unitcell.h"
-#include "source_pw/module_pwdft/parallel_grid.h"
+#include "source_base/parallel_grid.h"
 
 class Structure_Factor
 {
